refactor(blorktools): mesh info panel to widget
