+++ packages/blorktools/src/asset_debugger/util/scene/css3d-scene-controller.js
@@ -12,7 +12,7 @@ import {
     animationStack,
     resetAnimationState
 } from '../state/css3d-state';
-import { createMeshInfoPanel } from '../../panels/mesh-info-panel';
+import { createMeshInfoPanel } from '../../widgets/mesh-info-widget';
 import {     
     setAnimationCss3dRenderer,
     setAnimationCss3dScene,
