+++ packages/blorktools/src/asset_debugger/widgets/mesh-info-widget.js
